move TracingDirectives to metrics pkg
